drivers: usb: otg: add device tree support to otg library
Adds an API to get usb phy by passing a device node phandle value. Since
now it's possible to obtain phy by phandle, the checks in usb_add_phy
for a valid phy type is removed (now it's just a debug message if a user
tries to add a phy with undefined type).
This also allows to add multiple phys of same type.

Cc: Richard Zhao <richard.zhao@freescale.com>
Cc: Marek Vasut <marex@denx.de>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Ajay Kumar Gupta <ajay.gupta@ti.com>