ARM: OMAP2+: PM: clean up some debugfs functions
Move some debugfs functions into the powerdomain and clockdomain code;
the ultimate goal is to stop exporting some of the
powerdomain/clockdomain functions.  Get rid of the nasty hacks that
avoid executing some of the debug functions based on string
comparisons on powerdomain and clockdomain names.  Lock the
powerdomain for the duration of these functions so that the debug
information doesn't change during execution.  Also, while here, fix a
whitespace issue in mach-omap2/pm-debug.c.

Signed-off-by: Paul Walmsley <paul@pwsan.com>
Cc: Tero Kristo <t-kristo@ti.com>
Cc: Kevin Hilman <khilman@deeprootsystems.com>