mailbox: replace plat/cpu.h with corresponding soc.h
The AM33xx suspend resume functionality is built on top of the
common mailbox framework and cleanup series posted for review,
    http://www.spinics.net/lists/linux-omap/msg84035.html
but is breaking the build when rebased for v3.8-rc3 due to the
removal of soc.h from plat/cpu.h in commit,
	e8d3d47 ARM: OMAP2+: Drop plat/cpu.h for omap2plus

This is a quick temporary (though ugly) work-around to fix the
build for the time being. The cpu_is/soc_is macros should ideally
not be used by the drivers and the drivers should manage this
through either the platform data or device tree properties. The
current change is local to the mailbox driver code, which needs
a cleanup and reposting for the latest rc anyways.

Signed-off-by: Suman Anna <s-anna@ti.com>